+++ AAMPAnomalyMessageType.h
@@ -1,0 +1,34 @@
+/*
+ * If not stated otherwise in this file or this component's license file the
+ * following copyright and licenses apply:
+ *
+ * Copyright 2023 RDK Management
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+*/
+
+#ifndef AAMP_ANOMALY_MESSAGE_TYPE_H
+#define AAMP_ANOMALY_MESSAGE_TYPE_H
+
+/**
+ * @enum AAMPAnomalyMessageType
+ * @brief AAMP anomaly message types
+ */
+typedef enum
+{
+    ANOMALY_ERROR,    /**< Error Message */
+    ANOMALY_WARNING,  /**< Warning Message */
+    ANOMALY_TRACE      /**< Trace Message */
+} AAMPAnomalyMessageType;
+
+#endif // AAMP_ANOMALY_MESSAGE_TYPE_H
