+++ AuthTokenErrors.h
@@ -1,0 +1,32 @@
+/*
+ * If not stated otherwise in this file or this component's license file the
+ * following copyright and licenses apply:
+ *
+ * Copyright 2023 RDK Management
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+*/
+
+#ifndef AUTH_TOKEN_ERRORS_H
+#define AUTH_TOKEN_ERRORS_H
+
+/**
+ * @enum AuthTokenErrors
+ *  @brief Auth Token Failure codes
+ */
+enum AuthTokenErrors {
+    eAUTHTOKEN_TOKEN_PARSE_ERROR = -1,    /**< Auth token parse Error */
+    eAUTHTOKEN_INVALID_STATUS_CODE = -2   /**< Auth token Invalid status */
+};
+
+#endif // AUTH_TOKEN_ERRORS_H
