+++ drm/AampDRMLicManager.cpp
@@ -32,9 +32,8 @@
 #include "AampStreamSinkManager.h"
 #include "AampJsonObject.h"
 #include "AampConfig.h"
-
-
-
+#include "AAMPAnomalyMessageType.h"
+#include "AuthTokenErrors.h"
 
 #define SESSION_TOKEN_URL "http://localhost:50050/authService/getSessionToken"
