+++ main_aamp.h
@@ -56,7 +56,8 @@
 
 #include "LangCodePreference.h"
 #include "StreamOutputFormat.h"
-
+#include "VideoZoomMode.h"
+#include "StreamSink.h"
 
 /*! \mainpage
  *
@@ -68,657 +69,9 @@
 
 #define PrivAAMPState AAMPPlayerState // backwards compatibility for apps using native interface
 
-/**
- * @enum AAMPAnomalyMessageType
- * @brief AAMP anomaly message types
- */
-typedef enum
-{
-	ANOMALY_ERROR,    /**< Error Message */
-	ANOMALY_WARNING,  /**< Warning Message */
-	ANOMALY_TRACE	  /**< Trace Message */
-} AAMPAnomalyMessageType;
-
-/**
- * @struct TuneFailureMap
- * @brief  Structure holding aamp tune failure code and corresponding application error code and description
- */
-struct TuneFailureMap
-{
-	AAMPTuneFailure tuneFailure;    /**< Failure ID */
-	int code;                       /**< Error code */
-	const char* description;        /**< Textual description */
-};
-
-
-/**
- * @enum VideoZoomMode
- * @brief Video zoom mode
- */
-enum VideoZoomMode
-{
-	VIDEO_ZOOM_NONE,    /**< Video Zoom None */
-	VIDEO_ZOOM_DIRECT,  /**< Video Zoom Direct */
-	VIDEO_ZOOM_NORMAL,  /**< Video Zoom Normal */
-	VIDEO_ZOOM_16X9_STRETCH, /**< Video Zoom 16x9 stretch */
-	VIDEO_ZOOM_4x3_PILLAR_BOX, /**< Video Zoom 4x3 pillar box */
-	VIDEO_ZOOM_FULL, /**< Video Zoom Full */
-	VIDEO_ZOOM_GLOBAL/**< Video Zoom Global */
-};
-
-
+#if 0
 using AdObject = std::pair<std::string, std::string>;
-
-/**
- * @enum AuthTokenErrors
- *  @brief Auth Token Failure codes
- */
-enum AuthTokenErrors {
-	eAUTHTOKEN_TOKEN_PARSE_ERROR = -1,    /**< Auth token parse Error */
-	eAUTHTOKEN_INVALID_STATUS_CODE = -2   /**< Auth token Invalid status */
-};
-
-/**
- * @struct PreCacheUrlData
- * @brief Pre cache the data information
- */
-typedef struct PreCacheUrlData
-{
-	std::string url;
-	AampMediaType type;
-	PreCacheUrlData():url(""),type(eMEDIATYPE_VIDEO)
-	{
-	}
-} PreCacheUrlStruct;
-
-typedef std::vector < PreCacheUrlStruct> PreCacheUrlList;
-
-/**
- * @struct PlaybackQualityData
- * @brief Playback quality data information
- */
-typedef struct PlaybackQualityData
-{
-	long long rendered;
-	long long dropped;
-} PlaybackQualityStruct;
-
-/**
- * @struct AudioTrackInfo
- * @brief Structure for audio track information
- *        Holds information about an audio track in playlist
- */
-struct AudioTrackInfo
-{
-	std::string index;			/**< Index of track */
-	std::string language; 			/**< Language of track */
-	std::string rendition;			/**< role for DASH, group-id for HLS */
-	std::string name;			/**< Name of track info */
-	std::string codec;			/**< Codec of Audio track */
-	std::string characteristics;		/**< Characteristics field of audio track */
-	std::string label;			/**< label of audio track info */
-	int channels;				/**< number channels of track */
-	long bandwidth;				/**< Bandwidth value of track **/
-	int primaryKey; 			/**< used for ATSC to store key , this should not be exposed to app */
-	std::string contentType; 		/**< used for ATSC to propagate content type */
-	std::string mixType; 			/**< used for ATSC to propagate mix type */
-	std::string accessibilityType; 	 	/**< value of Accessibility */
-	bool isMuxed; 				/**< Flag to indicated muxed audio track ; this is used by AC4 tracks */
-	Accessibility accessibilityItem; 	/**< Field to store Accessibility Node */
-	std::string mType;			/**< Type field of track, to be populated by player */
-	bool isAvailable;
-	bool isDefault;
-
-	AudioTrackInfo() : index(), language(), rendition(), name(), codec(), characteristics(), channels(0),
-	bandwidth(0),primaryKey(0), contentType(), mixType(), accessibilityType(), isMuxed(false), label(), mType(), accessibilityItem(), isAvailable(true),
-	isDefault(false)
-	{
-	}
-
-	AudioTrackInfo(std::string idx, std::string lang, std::string rend, std::string trackName, std::string codecStr, std::string cha, int ch):
-		index(idx), language(lang), rendition(rend), name(trackName),
-		codec(codecStr), characteristics(cha), channels(ch), bandwidth(-1), primaryKey(0) , contentType(), mixType(), accessibilityType(), isMuxed(false), label(), mType(), accessibilityItem(),
-		isAvailable(true),isDefault(false)
-	{
-	}
-	AudioTrackInfo(std::string idx, std::string lang, std::string rend, std::string trackName, std::string codecStr, std::string cha, int ch,bool Default):
-		index(idx), language(lang), rendition(rend), name(trackName),
-		codec(codecStr), characteristics(cha), channels(ch), bandwidth(-1), primaryKey(0) , contentType(), mixType(), accessibilityType(), isMuxed(false), label(), mType(), accessibilityItem(),
-		isAvailable(true),isDefault(Default)
-	{
-	}
-
-	AudioTrackInfo(std::string idx, std::string lang, std::string rend, std::string trackName, std::string codecStr, int pk, std::string conType, std::string mixType):
-			index(idx), language(lang), rendition(rend), name(trackName),
-			codec(codecStr), characteristics(), channels(0), bandwidth(-1), primaryKey(pk),
-						contentType(conType), mixType(mixType), accessibilityType(), isMuxed(false), label(), mType(), accessibilityItem(),
-			isAvailable(true),isDefault(false)
-	{
-	}
-
-	AudioTrackInfo(std::string idx, std::string lang, std::string rend, std::string trackName, std::string codecStr, long bw, std::string typ, bool available):
-		index(idx), language(lang), rendition(rend), name(trackName),
-		codec(codecStr), characteristics(), channels(0), bandwidth(bw),primaryKey(0), contentType(), mixType(), accessibilityType(typ), isMuxed(false), label(), mType(), accessibilityItem(),
-		isAvailable(true),isDefault(false)
-	{
-	}
-
-	AudioTrackInfo(std::string idx, std::string lang, std::string rend, std::string trackName, std::string codecStr, long bw, int channel):
-		index(idx), language(lang), rendition(rend), name(trackName),
-		codec(codecStr), characteristics(), channels(channel), bandwidth(bw),primaryKey(0), contentType(), mixType(), accessibilityType(), isMuxed(false), label(), mType(), accessibilityItem(),
-		isAvailable(true),isDefault(false)
-	{
-	}
-
-	AudioTrackInfo(std::string idx, std::string lang, std::string rend, std::string trackName, std::string codecStr, long bw, int channel, bool muxed, bool available):
-		index(idx), language(lang), rendition(rend), name(trackName),
-		codec(codecStr), characteristics(), channels(channel), bandwidth(bw),primaryKey(0), contentType(), mixType(), accessibilityType(), isMuxed(muxed), label(), mType(), accessibilityItem(),
-		isAvailable(available),isDefault(false)
-	{
-	}
-
-	AudioTrackInfo(std::string idx, std::string lang, std::string rend, std::string trackName, std::string codecStr, long bw, std::string typ, bool muxed, std::string lab, std::string type, bool available):
-		index(idx), language(lang), rendition(rend), name(trackName),
-		codec(codecStr), characteristics(), channels(0), bandwidth(bw),primaryKey(0), contentType(), mixType(), accessibilityType(typ), isMuxed(muxed), label(lab), mType(type), accessibilityItem(),
-		isAvailable(available),isDefault(false)
-	{
-	}
-
-	AudioTrackInfo(std::string idx, std::string lang, std::string rend, std::string trackName, std::string codecStr, long bw, std::string typ, bool muxed, std::string lab, std::string type, Accessibility accessibility, bool available):
-		index(idx), language(lang), rendition(rend), name(trackName),
-		codec(codecStr), characteristics(), channels(0), bandwidth(bw),primaryKey(0), contentType(), mixType(), accessibilityType(typ), isMuxed(muxed), label(lab), mType(type), accessibilityItem(accessibility),
-		isAvailable(available),isDefault(false)
-	{
-	}
-
-	bool operator == (const AudioTrackInfo& track) const
-	{
-		return ((language == track.language) &&
-			(rendition == track.rendition) &&
-			(contentType == track.contentType) &&
-			(codec == track.codec) &&
-			(channels == track.channels) &&
-			(bandwidth == track.bandwidth) &&
-			(isMuxed == track.isMuxed) &&
-			(label == track.label) &&
-			(mType == track.mType) &&
-			(accessibilityItem == track.accessibilityItem) &&
-			(isDefault == track.isDefault));
-	}
-
-	bool operator < (const AudioTrackInfo& track) const
-	{
-		return (bandwidth < track.bandwidth);
-	}
-
-	bool operator > (const AudioTrackInfo& track) const
-	{
-		return (bandwidth > track.bandwidth);
-	}
-};
-
-/**
- * @brief Structure for text track information
- *        Holds information about a text track in playlist
- */
-struct TextTrackInfo
-{
-	std::string index;
-	std::string language;
-	bool isCC;
-	std::string rendition; //role for DASH, group-id for HLS
-	std::string name;
-	std::string instreamId;
-	std::string characteristics;
-	std::string codec;
-	std::string label; //Field for label
-	int primaryKey; // used for ATSC to store key , this should not be exposed to app.
-	std::string accessibilityType; //value of Accessibility
-	Accessibility accessibilityItem; /**< Field to store Accessibility Node */
-	std::string mType;
-	bool isAvailable;
-
-	//setter functions to reduce the number of parameters in constructor
-	void setLanguage(const std::string& lang) { language = lang; }
-	void setName(const std::string& trackName) { name = trackName; }
-	void setInstreamId(const std::string& id) { instreamId = id; }
-	void setCharacteristics(const std::string& cha) { characteristics = cha; }
-	void setCodec(const std::string& codecStr) { codec = codecStr; }
-	void setLabel(const std::string& lab) { label = lab; }
-	void setPrimaryKey(int pk) { primaryKey = pk; }
-	void setAccessibilityType(const std::string& accType) { accessibilityType = accType; }
-	void setAccessibilityItem(const Accessibility& acc) { accessibilityItem = acc; }
-	void setType(const std::string& type) { mType = type; }
-	void setAvailable(bool available) { isAvailable = available; }
-
-	TextTrackInfo() : index(), language(), isCC(false), rendition(), name(), instreamId(), characteristics(), codec(), primaryKey(0), accessibilityType(), label(), mType(), accessibilityItem(),
-			  isAvailable(true)
-	{
-	}
-
-	//2 parameter constructor
-	TextTrackInfo(bool cc, std::string rend):
-		isCC(cc), rendition(rend), index(), language(), name(), instreamId(), characteristics(), codec(), primaryKey(0), accessibilityType(), label(), mType(), accessibilityItem(), isAvailable(true)
-	{
-	}
-
-	//8 parameter constructor
-	TextTrackInfo(std::string idx, std::string lang, bool cc, std::string rend, std::string trackName, std::string id, std::string cha, int pk):
-		index(idx), language(lang), isCC(cc), rendition(rend),
-		name(trackName), instreamId(id), characteristics(cha),
-		codec(), primaryKey(pk), accessibilityType(), label(), mType(), accessibilityItem(), isAvailable(true)
-	{
-	}
-
-	//12 parameter full constructor
-	TextTrackInfo(std::string idx, std::string lang, bool cc, std::string rend, std::string trackName, std::string codecStr, std::string cha, std::string accType, std::string lab, std::string type, Accessibility acc, bool available):
-		index(idx), language(lang), isCC(cc), rendition(rend),
-		name(trackName), instreamId(), characteristics(cha),
-		codec(codecStr), primaryKey(0), accessibilityType(accType), label(lab), mType(type), accessibilityItem(acc), isAvailable(available)
-	{
-	}
-
-	void set (std::string idx, std::string lang, bool cc, std::string rend, std::string trackName, std::string codecStr, std::string cha,
-			std::string accType, std::string lab, std::string type, Accessibility acc)
-	{
-		index = idx;
-		language = lang;
-		isCC = cc;
-		rendition = rend;
-		name = trackName;
-		characteristics = cha;
-		codec = codecStr;
-		accessibilityType = accType;
-		label = lab;
-		accessibilityItem = acc;
-		mType = type;
-	}
-
-	bool operator == (const TextTrackInfo& track) const
-	{
-		return ((language == track.language) &&
-			(isCC == track.isCC) &&
-			(rendition == track.rendition) &&
-			(name == track.name) &&
-			(characteristics == track.characteristics) &&
-			(codec == track.codec) &&
-			(accessibilityType == track.accessibilityType) &&
-			(label == track.label) &&
-			(accessibilityItem == track.accessibilityItem) &&
-			(mType == track.mType));
-	}
-
-	bool operator < (const TextTrackInfo& track) const
-	{
-		return (index < track.index);
-	}
-
-	bool operator > (const TextTrackInfo& track) const
-	{
-		return (index > track.index);
-	}
-
-};
-
-/**
- * @class StreamSink
- * @brief GStreamer Abstraction class for the implementation of AAMPGstPlayer and gstaamp plugin
- */
-class StreamSink
-{
-public:
-
-	/**
-	 *   @brief  Configure output formats
-	 *
-	 *   @param[in]  format - Video output format.
-	 *   @param[in]  audioFormat - Audio output format.
-	 *   @param[in]  auxFormat - Aux audio output format.
-	 *   @param[in]  bESChangeStatus - Flag to keep force configure the pipeline value
-	 *   @param[in]  forwardAudioToAux - Flag denotes if audio buffers have to be forwarded to aux pipeline
-	 *   @param[in]  setReadyAfterPipelineCreation - Flag denotes if pipeline has to be reset to ready or not
-	 *   @return void
-	 */
-	virtual void Configure(StreamOutputFormat format, StreamOutputFormat audioFormat, StreamOutputFormat auxFormat, StreamOutputFormat subFormat, bool bESChangeStatus, bool forwardAudioToAux, bool setReadyAfterPipelineCreation=false){}
-
-	/**
-	 *   @brief  API to send audio/video buffer into the sink.
-	 *
-	 *   @param[in]  mediaType - Type of the media.
-	 *   @param[in]  ptr - Pointer to the buffer; caller responsible of freeing memory
-	 *   @param[in]  len - Buffer length.
-	 *   @param[in]  fpts - Presentation Time Stamp.
-	 *   @param[in]  fdts - Decode Time Stamp
-	 *   @param[in]  fDuration - Buffer duration.
-	 *   @return void
-	 */
-	virtual bool SendCopy( AampMediaType mediaType, const void *ptr, size_t len, double fpts, double fdts, double fDuration)= 0;
-
-	/**
-	 *   @brief  API to send audio/video buffer into the sink.
-	 *
-	 *   @param[in]  mediaType - Type of the media.
-	 *   @param[in]  buffer - Pointer to the AampGrowableBuffer; ownership is taken by the sink
-	 *   @param[in]  fpts - Presentation Time Stamp.
-	 *   @param[in]  fdts - Decode Time Stamp
-	 *   @param[in]  fDuration - Buffer duration.
-	 *   @param[in]  fragmentPTSoffset - Offset PTS
-	 *   @param[in]  initFragment - flag for buffer type (init, data)
-	 *   @return void
-	 */
-	virtual bool SendTransfer( AampMediaType mediaType, void *ptr, size_t len, double fpts, double fdts, double fDuration, double fragmentPTSoffset, bool initFragment = false, bool discontinuity = false)= 0;
-
-	/**
-	 *   @brief  Checks pipeline is configured for media type
-	 *
-	 *   @param[in]  mediaType - Media Type
-	 *   @return void
-	 */
-	virtual bool PipelineConfiguredForMedia(AampMediaType type){return true;}
-
-	/**
-	 *   @brief  Notifies EOS to sink
-	 *
-	 *   @param[in]  mediaType - Media Type
-	 *   @return void
-	 */
-	virtual void EndOfStreamReached(AampMediaType mediaType){}
-
-	/**
-	 *   @brief Start the stream
-	 *
-	 *   @return void
-	 */
-	virtual void Stream(void){}
-
-	/**
-	 *   @fn Stop
-	 *
-	 *   @param[in]  keepLastFrame - Keep the last frame on screen (true/false)
-	 *   @return void
-	 */
-	virtual void Stop(bool keepLastFrame){}
-
-	/**
-	 *   @brief Flush the pipeline
-	 *
-	 *   @param[in]  position - playback position
-	 *   @param[in]  rate - Speed
-	 *   @param[in]  shouldTearDown - if pipeline is not in a valid state, tear down pipeline
-	 *   @return void
-	 */
-	virtual void Flush(double position = 0, int rate = AAMP_NORMAL_PLAY_RATE, bool shouldTearDown = true){}
-
-	/**
-	 *   @brief Flush the audio playbin
-	 *   @param[in]  position - playback position
-	 *   @return void
-	 */
-	virtual void FlushTrack(AampMediaType mediaType,double position = 0){}
-
-	/**
-	 *   @brief Set player rate to audio/video sink
-	 *
-	 *   @param[in]  rate - Speed
-	 *   @return true if player rate is set successfully
-	 */
-	virtual bool SetPlayBackRate ( double rate ){return true;}
-	/**
-	 *   @brief Adjust the pipeline
-	 *
-	 *   @param[in]  position - playback position
-	 *   @param[in]  rate - Speed
-	 *   @return void
-	 */
-	virtual bool AdjustPlayBackRate(double position, double rate){ return true; }
-
-	/**
-	 *   @brief Enabled or disable playback pause
-	 *
-	 *   @param[in] pause  Enable/Disable
-	 *   @param[in] forceStopGstreamerPreBuffering - true for disabling buffer-in-progress
-	 *   @return true if content successfully paused
-	 */
-	virtual bool Pause(bool pause, bool forceStopGstreamerPreBuffering){ return true; }
-
-	/**
-	 *   @brief Get playback duration in milliseconds
-	 *
-	 *   @return duration in ms.
-	 */
-	virtual long GetDurationMilliseconds(void){ return 0; };
-
-	/**
-	 *   @brief Get playback position in milliseconds
-	 *
-	 *   @return Position in ms.
-	 */
-	virtual long long GetPositionMilliseconds(void){ return 0; };
-
-	/**
-	 *   @brief Get Video 90 KHz Video PTS
-	 *
-	 *   @return video PTS
-	 */
-	virtual long long GetVideoPTS(void){ return 2; };
-
-	/**
-	 *   @brief Get closed caption handle
-	 *
-	 *   @return Closed caption handle
-	 */
-	virtual unsigned long getCCDecoderHandle(void) { return 0; };
-
-	/**
-	 *   @brief Set video display rectangle co-ordinates
-	 *
-	 *   @param[in]  x - x position
-	 *   @param[in]  y - y position
-	 *   @param[in]  w - Width
-	 *   @param[in]  h - Height
-	 *   @return void
-	 */
-	virtual void SetVideoRectangle(int x, int y, int w, int h){};
-
-	/**
-	 *   @brief Set video zoom state
-	 *
-	 *   @param[in]  zoom - Zoom mode
-	 *   @return void
-	 */
-	virtual void SetVideoZoom(VideoZoomMode zoom){};
-
-	/**
-	 *   @brief Set video mute state
-	 *
-	 *   @param[in] muted - true: video muted, false: video unmuted
-	 *   @return void
-	 */
-	virtual void SetVideoMute(bool muted){};
-
-	/**
-	 *   @brief Set subtitle mute state
-	 *
-	 *   @param[in] muted - true: subtitle muted, false: subtitle unmuted
-	 *   @return void
-	 */
-	virtual void SetSubtitleMute(bool muted){};
-
-	/**
-	 *   @brief Set subtitle pts offset in sink
-	 *
-	 *   @param[in] pts_offset - pts offset for subs display
-	 *   @return void
-	 */
-	virtual void SetSubtitlePtsOffset(std::uint64_t pts_offset){};
-
-	/**
-	 *   @brief Set volume level
-	 *
-	 *   @param[in]  volume - Minimum 0, maximum 100.
-	 *   @return void
-	 */
-	virtual void SetAudioVolume(int volume){};
-
-	/**
-	 *   @brief StreamSink Dtor
-	 */
-	virtual ~StreamSink(){};
-
-	/**
-	 *   @brief Process PTS discontinuity for a stream type
-	 *
-	 *   @param[in]  mediaType - Media Type
-	 *   @return TRUE if discontinuity processed
-	 */
-	virtual bool Discontinuity( AampMediaType mediaType) = 0;
-
-
-	/**
-	 * @brief Check if PTS is changing
-	 *
-	 * @param[in] timeout - max time period within which PTS hasn't changed
-	 * @retval true if PTS is changing, false if PTS hasn't changed for timeout msecs
-	 */
-	virtual bool CheckForPTSChangeWithTimeout(long timeout) { return true; }
-
-	/**
-	 *   @brief Check whether cache is empty
-	 *
-	 *   @param[in]  mediaType - Media Type
-	 *   @return true: empty, false: not empty
-	 */
-	virtual bool IsCacheEmpty(AampMediaType mediaType){ return true; };
-
-	/**
-	 * @brief Reset EOS SignalledFlag
-	 */
-	virtual void ResetEOSSignalledFlag(){};
-
-	/**
-	 *   @brief API to notify that fragment caching done
-	 *
-	 *   @return void
-	 */
-	virtual void NotifyFragmentCachingComplete(){};
-
-	/**
-	 *   @brief API to notify that fragment caching is ongoing
-	 *
-	 *   @return void
-	 */
-	virtual void NotifyFragmentCachingOngoing(){};
-
-	/**
-	 *   @brief Get the video dimensions
-	 *
-	 *   @param[out]  w - Width
-	 *   @param[out]  h - Height
-	 *   @return void
-	 */
-	virtual void GetVideoSize(int &w, int &h){};
-
-	/**
-	 *   @brief Queue-up the protection event.
-	 *
-	 *   @param[in]  protSystemId - DRM system ID.
-	 *   @param[in]  ptr - Pointer to the protection data.
-	 *   @param[in]  len - Length of the protection data.
-	 *   @return void
-	 */
-	virtual void QueueProtectionEvent(const char *protSystemId, const void *ptr, size_t len, AampMediaType type) {};
-
-	/**
-	 *   @brief Clear the protection event.
-	 *
-	 *   @return void
-	 */
-	virtual void ClearProtectionEvent() {};
-
-	/**
-	 *   @brief Signal discontinuity on trickmode if restamping is done by stream sink.
-	 *
-	 *   @return void
-	 */
-	virtual void SignalTrickModeDiscontinuity() {};
-
-	/**
-	 *   @brief Seek stream sink to desired position and playback rate with a flushing seek
-	 *
-	 *   @param[in]  position - desired playback position.
-	 *   @param[in]  rate - desired playback rate.
-	 *   @return void
-	 */
-	virtual void SeekStreamSink(double position, double rate) {};
-
-	/**
-	 *   @brief Get the video window co-ordinates
-	 *
-	 *   @return current video co-ordinates in x,y,w,h format
-	 */
-	virtual std::string GetVideoRectangle() { return std::string(); };
-
-	/**
-	 *   @brief Stop buffering in sink
-	 *
-	 *   @param[in] forceStop - true if buffering to be stopped without any checks
-	 *   @return void
-	 */
-	virtual void StopBuffering(bool forceStop) { };
-
-	/**
-	 *   @brief Set the text style of the subtitle to the options passed
-	 *   @param[in] options - reference to the Json string that contains the information
-	 *   @return - true indicating successful operation in passing options to the parser
-	 */
-	virtual bool SetTextStyle(const std::string &options) { return false; }
-
-	/**
-	 *   @brief Get the set text style of the subtitle
-	 *   @param[in] textStyle - Json string that contains text style
-	 *   @return - true indicating indicating that subtitles are enabled
-	 */
-	virtual bool GetTextStyle(std::string &textStyle) { return false; }
-
-	/**
-	 * @brief API to set track Id to audio sync property in case of AC4 audio
-	 *
-	 * @param[in] trackId - AC4 track Id parsed by aamp based of preference
-	 * @return bool status of API
-	 */
-
-	/**
-	 *   @brief Get the video playback quality
-	 *
-	 *   @return current video playback quality
-	 */
-	virtual PlaybackQualityStruct* GetVideoPlaybackQuality() { return NULL; };
-
-	/**
-	 * @brief Signal the new clock to subtitle module
-	 * @return - true indicating successful operation in sending the clock update
-	 */
-	virtual bool SignalSubtitleClock( void ) { return false; };
-
-	/**
-	 * @fn SetPauseOnPlayback
-	 * @brief Set to pause on next playback start
-	 * @param[in] enable - Flag to set whether enabled
-	 */
-	virtual void SetPauseOnStartPlayback(bool enable) {};
-
-	/**
- 	* @brief Notifies the injector to resume buffer pushing.
- 	*/
-	virtual void NotifyInjectorToResume() {};
-
-	/**
- 	* @brief Notifies the injector to pause buffer pushing.
- 	*/
-	virtual void NotifyInjectorToPause() {};
-
-};
-
+#endif
 
 /**
  * @class PlayerInstanceAAMP
